+++ src/client/sound/Sound.cpp
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-//    Scorched3D (c) 2000-2011, 2024, 2025
+//    Scorched3D (c) 2000-2011, 2024, 2025, 2026
 //
 //    This file is part of Scorched3D.
 //
@@ -205,13 +205,18 @@ void Sound::showSoundBuffers()
 		PlayingSoundSource *source = (*itor);
 		if (source->getVirtualSource())
 		{
-			Logger::log(S3D::formatStringBuffer("%i - %u,%f - %s%s:%s",
-				i,
-				source->getVirtualSource()->getPriority(),
-				source->getVirtualSource()->getDistance(),
-				(source->getStopped()?"Finished":(source->getVirtualSource()->getPlaying()?"Playing":"Stopped")),
-				(source->getVirtualSource()->getLooping()?"(Looped)":""),
-				source->getVirtualSource()->getBuffer()->getFileName()));
+			Logger::log(
+				S3D::formatStringBuffer(
+					"%i - %u,%f - %s%s:%s",
+					i,
+					source->getVirtualSource()->getPriority(),
+					(double)(source->getVirtualSource()->getDistance()),
+					( source->getStopped() ? "Finished"
+										   : ( source->getVirtualSource()->getPlaying() ? "Playing" : "Stopped" ) ),
+					( source->getVirtualSource()->getLooping() ? "(Looped)" : "" ),
+					source->getVirtualSource()->getBuffer()->getFileName()
+				)
+			);
 		}
 		else
 		{
